Fix for segfault for big matrices sizes.

+++ algorithms/bini322_10_52_approx.hpp
@@ -19,7 +19,7 @@ void S_Add1(Matrix<Scalar>& S1, Matrix<Scalar>& S2, Matrix<Scalar>& C, double x,
 #endif
     for (int j = 0; j < C.n(); ++j) {
         for (int i = 0; i < C.m(); ++i) {
-            dataC[i + j * strideC] = dataS1[i + j * strideS1] + dataS2[i + j * strideS2];
+            dataC[i + static_cast<size_t>(j) * strideC] = dataS1[i + static_cast<size_t>(j) * strideS1] + dataS2[i + static_cast<size_t>(j) * strideS2];
         }
     }
 }
@@ -35,7 +35,7 @@ void S_Add2(Matrix<Scalar>& S1, Matrix<Scalar>& C, double x, bool sequential) {
 #endif
     for (int j = 0; j < C.n(); ++j) {
         for (int i = 0; i < C.m(); ++i) {
-            dataC[i + j * strideC] = dataS1[i + j * strideS1];
+            dataC[i + static_cast<size_t>(j) * strideC] = dataS1[i + static_cast<size_t>(j) * strideS1];
         }
     }
 }
@@ -51,7 +51,7 @@ void S_Add3(Matrix<Scalar>& S1, Matrix<Scalar>& C, double x, bool sequential) {
 #endif
     for (int j = 0; j < C.n(); ++j) {
         for (int i = 0; i < C.m(); ++i) {
-            dataC[i + j * strideC] = dataS1[i + j * strideS1];
+            dataC[i + static_cast<size_t>(j) * strideC] = dataS1[i + static_cast<size_t>(j) * strideS1];
         }
     }
 }
@@ -69,7 +69,7 @@ void S_Add4(Matrix<Scalar>& S1, Matrix<Scalar>& S2, Matrix<Scalar>& C, double x,
 #endif
     for (int j = 0; j < C.n(); ++j) {
         for (int i = 0; i < C.m(); ++i) {
-            dataC[i + j * strideC] = Scalar(x) * dataS1[i + j * strideS1] + dataS2[i + j * strideS2];
+            dataC[i + static_cast<size_t>(j) * strideC] = Scalar(x) * dataS1[i + static_cast<size_t>(j) * strideS1] + dataS2[i + static_cast<size_t>(j) * strideS2];
         }
     }
 }
@@ -87,7 +87,7 @@ void S_Add5(Matrix<Scalar>& S1, Matrix<Scalar>& S2, Matrix<Scalar>& C, double x,
 #endif
     for (int j = 0; j < C.n(); ++j) {
         for (int i = 0; i < C.m(); ++i) {
-            dataC[i + j * strideC] = dataS1[i + j * strideS1] + Scalar(x) * dataS2[i + j * strideS2];
+            dataC[i + static_cast<size_t>(j) * strideC] = dataS1[i + static_cast<size_t>(j) * strideS1] + Scalar(x) * dataS2[i + static_cast<size_t>(j) * strideS2];
         }
     }
 }
@@ -105,7 +105,7 @@ void S_Add6(Matrix<Scalar>& S1, Matrix<Scalar>& S2, Matrix<Scalar>& C, double x,
 #endif
     for (int j = 0; j < C.n(); ++j) {
         for (int i = 0; i < C.m(); ++i) {
-            dataC[i + j * strideC] = dataS1[i + j * strideS1] + dataS2[i + j * strideS2];
+            dataC[i + static_cast<size_t>(j) * strideC] = dataS1[i + static_cast<size_t>(j) * strideS1] + dataS2[i + static_cast<size_t>(j) * strideS2];
         }
     }
 }
@@ -121,7 +121,7 @@ void S_Add7(Matrix<Scalar>& S1, Matrix<Scalar>& C, double x, bool sequential) {
 #endif
     for (int j = 0; j < C.n(); ++j) {
         for (int i = 0; i < C.m(); ++i) {
-            dataC[i + j * strideC] = dataS1[i + j * strideS1];
+            dataC[i + static_cast<size_t>(j) * strideC] = dataS1[i + static_cast<size_t>(j) * strideS1];
         }
     }
 }
@@ -137,7 +137,7 @@ void S_Add8(Matrix<Scalar>& S1, Matrix<Scalar>& C, double x, bool sequential) {
 #endif
     for (int j = 0; j < C.n(); ++j) {
         for (int i = 0; i < C.m(); ++i) {
-            dataC[i + j * strideC] = dataS1[i + j * strideS1];
+            dataC[i + static_cast<size_t>(j) * strideC] = dataS1[i + static_cast<size_t>(j) * strideS1];
         }
     }
 }
@@ -155,7 +155,7 @@ void S_Add9(Matrix<Scalar>& S1, Matrix<Scalar>& S2, Matrix<Scalar>& C, double x,
 #endif
     for (int j = 0; j < C.n(); ++j) {
         for (int i = 0; i < C.m(); ++i) {
-            dataC[i + j * strideC] = Scalar(x) * dataS1[i + j * strideS1] + dataS2[i + j * strideS2];
+            dataC[i + static_cast<size_t>(j) * strideC] = Scalar(x) * dataS1[i + static_cast<size_t>(j) * strideS1] + dataS2[i + static_cast<size_t>(j) * strideS2];
         }
     }
 }
@@ -173,7 +173,7 @@ void S_Add10(Matrix<Scalar>& S1, Matrix<Scalar>& S2, Matrix<Scalar>& C, double x
 #endif
     for (int j = 0; j < C.n(); ++j) {
         for (int i = 0; i < C.m(); ++i) {
-            dataC[i + j * strideC] = dataS1[i + j * strideS1] + Scalar(x) * dataS2[i + j * strideS2];
+            dataC[i + static_cast<size_t>(j) * strideC] = dataS1[i + static_cast<size_t>(j) * strideS1] + Scalar(x) * dataS2[i + static_cast<size_t>(j) * strideS2];
         }
     }
 }
@@ -191,7 +191,7 @@ void T_Add1(Matrix<Scalar>& T1, Matrix<Scalar>& T2, Matrix<Scalar>& C, double x,
 #endif
     for (int j = 0; j < C.n(); ++j) {
         for (int i = 0; i < C.m(); ++i) {
-            dataC[i + j * strideC] = Scalar(x) * dataT1[i + j * strideT1] + dataT2[i + j * strideT2];
+            dataC[i + static_cast<size_t>(j) * strideC] = Scalar(x) * dataT1[i + static_cast<size_t>(j) * strideT1] + dataT2[i + static_cast<size_t>(j) * strideT2];
         }
     }
 }
@@ -209,7 +209,7 @@ void T_Add2(Matrix<Scalar>& T1, Matrix<Scalar>& T2, Matrix<Scalar>& C, double x,
 #endif
     for (int j = 0; j < C.n(); ++j) {
         for (int i = 0; i < C.m(); ++i) {
-            dataC[i + j * strideC] = -dataT1[i + j * strideT1] -dataT2[i + j * strideT2];
+            dataC[i + static_cast<size_t>(j) * strideC] = -dataT1[i + static_cast<size_t>(j) * strideT1] -dataT2[i + static_cast<size_t>(j) * strideT2];
         }
     }
 }
@@ -225,7 +225,7 @@ void T_Add3(Matrix<Scalar>& T1, Matrix<Scalar>& C, double x, bool sequential) {
 #endif
     for (int j = 0; j < C.n(); ++j) {
         for (int i = 0; i < C.m(); ++i) {
-            dataC[i + j * strideC] = dataT1[i + j * strideT1];
+            dataC[i + static_cast<size_t>(j) * strideC] = dataT1[i + static_cast<size_t>(j) * strideT1];
         }
     }
 }
@@ -243,7 +243,7 @@ void T_Add4(Matrix<Scalar>& T1, Matrix<Scalar>& T2, Matrix<Scalar>& C, double x,
 #endif
     for (int j = 0; j < C.n(); ++j) {
         for (int i = 0; i < C.m(); ++i) {
-            dataC[i + j * strideC] = Scalar(-(x)) * dataT1[i + j * strideT1] + dataT2[i + j * strideT2];
+            dataC[i + static_cast<size_t>(j) * strideC] = Scalar(-(x)) * dataT1[i + static_cast<size_t>(j) * strideT1] + dataT2[i + static_cast<size_t>(j) * strideT2];
         }
     }
 }
@@ -261,7 +261,7 @@ void T_Add5(Matrix<Scalar>& T1, Matrix<Scalar>& T2, Matrix<Scalar>& C, double x,
 #endif
     for (int j = 0; j < C.n(); ++j) {
         for (int i = 0; i < C.m(); ++i) {
-            dataC[i + j * strideC] = Scalar(x) * dataT1[i + j * strideT1] + dataT2[i + j * strideT2];
+            dataC[i + static_cast<size_t>(j) * strideC] = Scalar(x) * dataT1[i + static_cast<size_t>(j) * strideT1] + dataT2[i + static_cast<size_t>(j) * strideT2];
         }
     }
 }
@@ -279,7 +279,7 @@ void T_Add6(Matrix<Scalar>& T1, Matrix<Scalar>& T2, Matrix<Scalar>& C, double x,
 #endif
     for (int j = 0; j < C.n(); ++j) {
         for (int i = 0; i < C.m(); ++i) {
-            dataC[i + j * strideC] = dataT1[i + j * strideT1] + Scalar(x) * dataT2[i + j * strideT2];
+            dataC[i + static_cast<size_t>(j) * strideC] = dataT1[i + static_cast<size_t>(j) * strideT1] + Scalar(x) * dataT2[i + static_cast<size_t>(j) * strideT2];
         }
     }
 }
@@ -295,7 +295,7 @@ void T_Add7(Matrix<Scalar>& T1, Matrix<Scalar>& C, double x, bool sequential) {
 #endif
     for (int j = 0; j < C.n(); ++j) {
         for (int i = 0; i < C.m(); ++i) {
-            dataC[i + j * strideC] = dataT1[i + j * strideT1];
+            dataC[i + static_cast<size_t>(j) * strideC] = dataT1[i + static_cast<size_t>(j) * strideT1];
         }
     }
 }
@@ -313,7 +313,7 @@ void T_Add8(Matrix<Scalar>& T1, Matrix<Scalar>& T2, Matrix<Scalar>& C, double x,
 #endif
     for (int j = 0; j < C.n(); ++j) {
         for (int i = 0; i < C.m(); ++i) {
-            dataC[i + j * strideC] = -dataT1[i + j * strideT1] -dataT2[i + j * strideT2];
+            dataC[i + static_cast<size_t>(j) * strideC] = -dataT1[i + static_cast<size_t>(j) * strideT1] -dataT2[i + static_cast<size_t>(j) * strideT2];
         }
     }
 }
@@ -331,7 +331,7 @@ void T_Add9(Matrix<Scalar>& T1, Matrix<Scalar>& T2, Matrix<Scalar>& C, double x,
 #endif
     for (int j = 0; j < C.n(); ++j) {
         for (int i = 0; i < C.m(); ++i) {
-            dataC[i + j * strideC] = dataT1[i + j * strideT1] + Scalar(x) * dataT2[i + j * strideT2];
+            dataC[i + static_cast<size_t>(j) * strideC] = dataT1[i + static_cast<size_t>(j) * strideT1] + Scalar(x) * dataT2[i + static_cast<size_t>(j) * strideT2];
         }
     }
 }
@@ -349,7 +349,7 @@ void T_Add10(Matrix<Scalar>& T1, Matrix<Scalar>& T2, Matrix<Scalar>& C, double x
 #endif
     for (int j = 0; j < C.n(); ++j) {
         for (int i = 0; i < C.m(); ++i) {
-            dataC[i + j * strideC] = dataT1[i + j * strideT1] + Scalar(-(x)) * dataT2[i + j * strideT2];
+            dataC[i + static_cast<size_t>(j) * strideC] = dataT1[i + static_cast<size_t>(j) * strideT1] + Scalar(-(x)) * dataT2[i + static_cast<size_t>(j) * strideT2];
         }
     }
 }
@@ -372,7 +372,7 @@ void M_Add1(Matrix<Scalar>& M1, Matrix<Scalar>& M2, Matrix<Scalar>& M3, Matrix<S
 #endif
         for (int j = 0; j < C.n(); ++j) {
             for (int i = 0; i < C.m(); ++i) {
-                dataC[i + j * strideC] = Scalar(1.0 / (x)) * dataM1[i + j * strideM1] + Scalar(1.0 / (x)) * dataM2[i + j * strideM2] + Scalar(-(1.0 / (x))) * dataM3[i + j * strideM3] + Scalar(1.0 / (x)) * dataM4[i + j * strideM4] + beta * dataC[i + j * strideC];
+                dataC[i + static_cast<size_t>(j) * strideC] = Scalar(1.0 / (x)) * dataM1[i + static_cast<size_t>(j) * strideM1] + Scalar(1.0 / (x)) * dataM2[i + static_cast<size_t>(j) * strideM2] + Scalar(-(1.0 / (x))) * dataM3[i + static_cast<size_t>(j) * strideM3] + Scalar(1.0 / (x)) * dataM4[i + static_cast<size_t>(j) * strideM4] + beta * dataC[i + static_cast<size_t>(j) * strideC];
             }
         }
     } else {
@@ -381,7 +381,7 @@ void M_Add1(Matrix<Scalar>& M1, Matrix<Scalar>& M2, Matrix<Scalar>& M3, Matrix<S
 #endif
         for (int j = 0; j < C.n(); ++j) {
             for (int i = 0; i < C.m(); ++i) {
-                dataC[i + j * strideC] = Scalar(1.0 / (x)) * dataM1[i + j * strideM1] + Scalar(1.0 / (x)) * dataM2[i + j * strideM2] + Scalar(-(1.0 / (x))) * dataM3[i + j * strideM3] + Scalar(1.0 / (x)) * dataM4[i + j * strideM4];
+                dataC[i + static_cast<size_t>(j) * strideC] = Scalar(1.0 / (x)) * dataM1[i + static_cast<size_t>(j) * strideM1] + Scalar(1.0 / (x)) * dataM2[i + static_cast<size_t>(j) * strideM2] + Scalar(-(1.0 / (x))) * dataM3[i + static_cast<size_t>(j) * strideM3] + Scalar(1.0 / (x)) * dataM4[i + static_cast<size_t>(j) * strideM4];
             }
         }
     }
@@ -401,7 +401,7 @@ void M_Add2(Matrix<Scalar>& M1, Matrix<Scalar>& M2, Matrix<Scalar>& C, double x,
 #endif
         for (int j = 0; j < C.n(); ++j) {
             for (int i = 0; i < C.m(); ++i) {
-                dataC[i + j * strideC] = Scalar(-(1.0 / (x))) * dataM1[i + j * strideM1] + Scalar(1.0 / (x)) * dataM2[i + j * strideM2] + beta * dataC[i + j * strideC];
+                dataC[i + static_cast<size_t>(j) * strideC] = Scalar(-(1.0 / (x))) * dataM1[i + static_cast<size_t>(j) * strideM1] + Scalar(1.0 / (x)) * dataM2[i + static_cast<size_t>(j) * strideM2] + beta * dataC[i + static_cast<size_t>(j) * strideC];
             }
         }
     } else {
@@ -410,7 +410,7 @@ void M_Add2(Matrix<Scalar>& M1, Matrix<Scalar>& M2, Matrix<Scalar>& C, double x,
 #endif
         for (int j = 0; j < C.n(); ++j) {
             for (int i = 0; i < C.m(); ++i) {
-                dataC[i + j * strideC] = Scalar(-(1.0 / (x))) * dataM1[i + j * strideM1] + Scalar(1.0 / (x)) * dataM2[i + j * strideM2];
+                dataC[i + static_cast<size_t>(j) * strideC] = Scalar(-(1.0 / (x))) * dataM1[i + static_cast<size_t>(j) * strideM1] + Scalar(1.0 / (x)) * dataM2[i + static_cast<size_t>(j) * strideM2];
             }
         }
     }
@@ -432,7 +432,7 @@ void M_Add3(Matrix<Scalar>& M1, Matrix<Scalar>& M2, Matrix<Scalar>& M3, Matrix<S
 #endif
         for (int j = 0; j < C.n(); ++j) {
             for (int i = 0; i < C.m(); ++i) {
-                dataC[i + j * strideC] = dataM1[i + j * strideM1] + dataM2[i + j * strideM2] -dataM3[i + j * strideM3] + beta * dataC[i + j * strideC];
+                dataC[i + static_cast<size_t>(j) * strideC] = dataM1[i + static_cast<size_t>(j) * strideM1] + dataM2[i + static_cast<size_t>(j) * strideM2] -dataM3[i + static_cast<size_t>(j) * strideM3] + beta * dataC[i + static_cast<size_t>(j) * strideC];
             }
         }
     } else {
@@ -441,7 +441,7 @@ void M_Add3(Matrix<Scalar>& M1, Matrix<Scalar>& M2, Matrix<Scalar>& M3, Matrix<S
 #endif
         for (int j = 0; j < C.n(); ++j) {
             for (int i = 0; i < C.m(); ++i) {
-                dataC[i + j * strideC] = dataM1[i + j * strideM1] + dataM2[i + j * strideM2] -dataM3[i + j * strideM3];
+                dataC[i + static_cast<size_t>(j) * strideC] = dataM1[i + static_cast<size_t>(j) * strideM1] + dataM2[i + static_cast<size_t>(j) * strideM2] -dataM3[i + static_cast<size_t>(j) * strideM3];
             }
         }
     }
@@ -463,7 +463,7 @@ void M_Add4(Matrix<Scalar>& M1, Matrix<Scalar>& M2, Matrix<Scalar>& M3, Matrix<S
 #endif
         for (int j = 0; j < C.n(); ++j) {
             for (int i = 0; i < C.m(); ++i) {
-                dataC[i + j * strideC] = dataM1[i + j * strideM1] -dataM2[i + j * strideM2] + dataM3[i + j * strideM3] + beta * dataC[i + j * strideC];
+                dataC[i + static_cast<size_t>(j) * strideC] = dataM1[i + static_cast<size_t>(j) * strideM1] -dataM2[i + static_cast<size_t>(j) * strideM2] + dataM3[i + static_cast<size_t>(j) * strideM3] + beta * dataC[i + static_cast<size_t>(j) * strideC];
             }
         }
     } else {
@@ -472,7 +472,7 @@ void M_Add4(Matrix<Scalar>& M1, Matrix<Scalar>& M2, Matrix<Scalar>& M3, Matrix<S
 #endif
         for (int j = 0; j < C.n(); ++j) {
             for (int i = 0; i < C.m(); ++i) {
-                dataC[i + j * strideC] = dataM1[i + j * strideM1] -dataM2[i + j * strideM2] + dataM3[i + j * strideM3];
+                dataC[i + static_cast<size_t>(j) * strideC] = dataM1[i + static_cast<size_t>(j) * strideM1] -dataM2[i + static_cast<size_t>(j) * strideM2] + dataM3[i + static_cast<size_t>(j) * strideM3];
             }
         }
     }
@@ -492,7 +492,7 @@ void M_Add5(Matrix<Scalar>& M1, Matrix<Scalar>& M2, Matrix<Scalar>& C, double x,
 #endif
         for (int j = 0; j < C.n(); ++j) {
             for (int i = 0; i < C.m(); ++i) {
-                dataC[i + j * strideC] = Scalar(-(1.0 / (x))) * dataM1[i + j * strideM1] + Scalar(1.0 / (x)) * dataM2[i + j * strideM2] + beta * dataC[i + j * strideC];
+                dataC[i + static_cast<size_t>(j) * strideC] = Scalar(-(1.0 / (x))) * dataM1[i + static_cast<size_t>(j) * strideM1] + Scalar(1.0 / (x)) * dataM2[i + static_cast<size_t>(j) * strideM2] + beta * dataC[i + static_cast<size_t>(j) * strideC];
             }
         }
     } else {
@@ -501,7 +501,7 @@ void M_Add5(Matrix<Scalar>& M1, Matrix<Scalar>& M2, Matrix<Scalar>& C, double x,
 #endif
         for (int j = 0; j < C.n(); ++j) {
             for (int i = 0; i < C.m(); ++i) {
-                dataC[i + j * strideC] = Scalar(-(1.0 / (x))) * dataM1[i + j * strideM1] + Scalar(1.0 / (x)) * dataM2[i + j * strideM2];
+                dataC[i + static_cast<size_t>(j) * strideC] = Scalar(-(1.0 / (x))) * dataM1[i + static_cast<size_t>(j) * strideM1] + Scalar(1.0 / (x)) * dataM2[i + static_cast<size_t>(j) * strideM2];
             }
         }
     }
@@ -525,7 +525,7 @@ void M_Add6(Matrix<Scalar>& M1, Matrix<Scalar>& M2, Matrix<Scalar>& M3, Matrix<S
 #endif
         for (int j = 0; j < C.n(); ++j) {
             for (int i = 0; i < C.m(); ++i) {
-                dataC[i + j * strideC] = Scalar(1.0 / (x)) * dataM1[i + j * strideM1] + Scalar(-(1.0 / (x))) * dataM2[i + j * strideM2] + Scalar(1.0 / (x)) * dataM3[i + j * strideM3] + Scalar(1.0 / (x)) * dataM4[i + j * strideM4] + beta * dataC[i + j * strideC];
+                dataC[i + static_cast<size_t>(j) * strideC] = Scalar(1.0 / (x)) * dataM1[i + static_cast<size_t>(j) * strideM1] + Scalar(-(1.0 / (x))) * dataM2[i + static_cast<size_t>(j) * strideM2] + Scalar(1.0 / (x)) * dataM3[i + static_cast<size_t>(j) * strideM3] + Scalar(1.0 / (x)) * dataM4[i + static_cast<size_t>(j) * strideM4] + beta * dataC[i + static_cast<size_t>(j) * strideC];
             }
         }
     } else {
@@ -534,7 +534,7 @@ void M_Add6(Matrix<Scalar>& M1, Matrix<Scalar>& M2, Matrix<Scalar>& M3, Matrix<S
 #endif
         for (int j = 0; j < C.n(); ++j) {
             for (int i = 0; i < C.m(); ++i) {
-                dataC[i + j * strideC] = Scalar(1.0 / (x)) * dataM1[i + j * strideM1] + Scalar(-(1.0 / (x))) * dataM2[i + j * strideM2] + Scalar(1.0 / (x)) * dataM3[i + j * strideM3] + Scalar(1.0 / (x)) * dataM4[i + j * strideM4];
+                dataC[i + static_cast<size_t>(j) * strideC] = Scalar(1.0 / (x)) * dataM1[i + static_cast<size_t>(j) * strideM1] + Scalar(-(1.0 / (x))) * dataM2[i + static_cast<size_t>(j) * strideM2] + Scalar(1.0 / (x)) * dataM3[i + static_cast<size_t>(j) * strideM3] + Scalar(1.0 / (x)) * dataM4[i + static_cast<size_t>(j) * strideM4];
             }
         }
     }
