+++ algorithms/classical222.hpp
@@ -17,7 +17,7 @@ void S_Add1(Matrix<Scalar>& S1, Matrix<Scalar>& C, double x, bool sequential) {
 #endif
     for (int j = 0; j < C.n(); ++j) {
         for (int i = 0; i < C.m(); ++i) {
-            dataC[i + j * strideC] = dataS1[i + j * strideS1];
+            dataC[i + static_cast<size_t>(j) * strideC] = dataS1[i + static_cast<size_t>(j) * strideS1];
         }
     }
 }
@@ -33,7 +33,7 @@ void S_Add2(Matrix<Scalar>& S1, Matrix<Scalar>& C, double x, bool sequential) {
 #endif
     for (int j = 0; j < C.n(); ++j) {
         for (int i = 0; i < C.m(); ++i) {
-            dataC[i + j * strideC] = dataS1[i + j * strideS1];
+            dataC[i + static_cast<size_t>(j) * strideC] = dataS1[i + static_cast<size_t>(j) * strideS1];
         }
     }
 }
@@ -49,7 +49,7 @@ void S_Add3(Matrix<Scalar>& S1, Matrix<Scalar>& C, double x, bool sequential) {
 #endif
     for (int j = 0; j < C.n(); ++j) {
         for (int i = 0; i < C.m(); ++i) {
-            dataC[i + j * strideC] = dataS1[i + j * strideS1];
+            dataC[i + static_cast<size_t>(j) * strideC] = dataS1[i + static_cast<size_t>(j) * strideS1];
         }
     }
 }
@@ -65,7 +65,7 @@ void S_Add4(Matrix<Scalar>& S1, Matrix<Scalar>& C, double x, bool sequential) {
 #endif
     for (int j = 0; j < C.n(); ++j) {
         for (int i = 0; i < C.m(); ++i) {
-            dataC[i + j * strideC] = dataS1[i + j * strideS1];
+            dataC[i + static_cast<size_t>(j) * strideC] = dataS1[i + static_cast<size_t>(j) * strideS1];
         }
     }
 }
@@ -81,7 +81,7 @@ void S_Add5(Matrix<Scalar>& S1, Matrix<Scalar>& C, double x, bool sequential) {
 #endif
     for (int j = 0; j < C.n(); ++j) {
         for (int i = 0; i < C.m(); ++i) {
-            dataC[i + j * strideC] = dataS1[i + j * strideS1];
+            dataC[i + static_cast<size_t>(j) * strideC] = dataS1[i + static_cast<size_t>(j) * strideS1];
         }
     }
 }
@@ -97,7 +97,7 @@ void S_Add6(Matrix<Scalar>& S1, Matrix<Scalar>& C, double x, bool sequential) {
 #endif
     for (int j = 0; j < C.n(); ++j) {
         for (int i = 0; i < C.m(); ++i) {
-            dataC[i + j * strideC] = dataS1[i + j * strideS1];
+            dataC[i + static_cast<size_t>(j) * strideC] = dataS1[i + static_cast<size_t>(j) * strideS1];
         }
     }
 }
@@ -113,7 +113,7 @@ void S_Add7(Matrix<Scalar>& S1, Matrix<Scalar>& C, double x, bool sequential) {
 #endif
     for (int j = 0; j < C.n(); ++j) {
         for (int i = 0; i < C.m(); ++i) {
-            dataC[i + j * strideC] = dataS1[i + j * strideS1];
+            dataC[i + static_cast<size_t>(j) * strideC] = dataS1[i + static_cast<size_t>(j) * strideS1];
         }
     }
 }
@@ -129,7 +129,7 @@ void S_Add8(Matrix<Scalar>& S1, Matrix<Scalar>& C, double x, bool sequential) {
 #endif
     for (int j = 0; j < C.n(); ++j) {
         for (int i = 0; i < C.m(); ++i) {
-            dataC[i + j * strideC] = dataS1[i + j * strideS1];
+            dataC[i + static_cast<size_t>(j) * strideC] = dataS1[i + static_cast<size_t>(j) * strideS1];
         }
     }
 }
@@ -145,7 +145,7 @@ void T_Add1(Matrix<Scalar>& T1, Matrix<Scalar>& C, double x, bool sequential) {
 #endif
     for (int j = 0; j < C.n(); ++j) {
         for (int i = 0; i < C.m(); ++i) {
-            dataC[i + j * strideC] = dataT1[i + j * strideT1];
+            dataC[i + static_cast<size_t>(j) * strideC] = dataT1[i + static_cast<size_t>(j) * strideT1];
         }
     }
 }
@@ -161,7 +161,7 @@ void T_Add2(Matrix<Scalar>& T1, Matrix<Scalar>& C, double x, bool sequential) {
 #endif
     for (int j = 0; j < C.n(); ++j) {
         for (int i = 0; i < C.m(); ++i) {
-            dataC[i + j * strideC] = dataT1[i + j * strideT1];
+            dataC[i + static_cast<size_t>(j) * strideC] = dataT1[i + static_cast<size_t>(j) * strideT1];
         }
     }
 }
@@ -177,7 +177,7 @@ void T_Add3(Matrix<Scalar>& T1, Matrix<Scalar>& C, double x, bool sequential) {
 #endif
     for (int j = 0; j < C.n(); ++j) {
         for (int i = 0; i < C.m(); ++i) {
-            dataC[i + j * strideC] = dataT1[i + j * strideT1];
+            dataC[i + static_cast<size_t>(j) * strideC] = dataT1[i + static_cast<size_t>(j) * strideT1];
         }
     }
 }
@@ -193,7 +193,7 @@ void T_Add4(Matrix<Scalar>& T1, Matrix<Scalar>& C, double x, bool sequential) {
 #endif
     for (int j = 0; j < C.n(); ++j) {
         for (int i = 0; i < C.m(); ++i) {
-            dataC[i + j * strideC] = dataT1[i + j * strideT1];
+            dataC[i + static_cast<size_t>(j) * strideC] = dataT1[i + static_cast<size_t>(j) * strideT1];
         }
     }
 }
@@ -209,7 +209,7 @@ void T_Add5(Matrix<Scalar>& T1, Matrix<Scalar>& C, double x, bool sequential) {
 #endif
     for (int j = 0; j < C.n(); ++j) {
         for (int i = 0; i < C.m(); ++i) {
-            dataC[i + j * strideC] = dataT1[i + j * strideT1];
+            dataC[i + static_cast<size_t>(j) * strideC] = dataT1[i + static_cast<size_t>(j) * strideT1];
         }
     }
 }
@@ -225,7 +225,7 @@ void T_Add6(Matrix<Scalar>& T1, Matrix<Scalar>& C, double x, bool sequential) {
 #endif
     for (int j = 0; j < C.n(); ++j) {
         for (int i = 0; i < C.m(); ++i) {
-            dataC[i + j * strideC] = dataT1[i + j * strideT1];
+            dataC[i + static_cast<size_t>(j) * strideC] = dataT1[i + static_cast<size_t>(j) * strideT1];
         }
     }
 }
@@ -241,7 +241,7 @@ void T_Add7(Matrix<Scalar>& T1, Matrix<Scalar>& C, double x, bool sequential) {
 #endif
     for (int j = 0; j < C.n(); ++j) {
         for (int i = 0; i < C.m(); ++i) {
-            dataC[i + j * strideC] = dataT1[i + j * strideT1];
+            dataC[i + static_cast<size_t>(j) * strideC] = dataT1[i + static_cast<size_t>(j) * strideT1];
         }
     }
 }
@@ -257,7 +257,7 @@ void T_Add8(Matrix<Scalar>& T1, Matrix<Scalar>& C, double x, bool sequential) {
 #endif
     for (int j = 0; j < C.n(); ++j) {
         for (int i = 0; i < C.m(); ++i) {
-            dataC[i + j * strideC] = dataT1[i + j * strideT1];
+            dataC[i + static_cast<size_t>(j) * strideC] = dataT1[i + static_cast<size_t>(j) * strideT1];
         }
     }
 }
@@ -276,7 +276,7 @@ void M_Add1(Matrix<Scalar>& M1, Matrix<Scalar>& M2, Matrix<Scalar>& C, double x,
 #endif
         for (int j = 0; j < C.n(); ++j) {
             for (int i = 0; i < C.m(); ++i) {
-                dataC[i + j * strideC] = dataM1[i + j * strideM1] + dataM2[i + j * strideM2] + beta * dataC[i + j * strideC];
+                dataC[i + static_cast<size_t>(j) * strideC] = dataM1[i + static_cast<size_t>(j) * strideM1] + dataM2[i + static_cast<size_t>(j) * strideM2] + beta * dataC[i + static_cast<size_t>(j) * strideC];
             }
         }
     } else {
@@ -285,7 +285,7 @@ void M_Add1(Matrix<Scalar>& M1, Matrix<Scalar>& M2, Matrix<Scalar>& C, double x,
 #endif
         for (int j = 0; j < C.n(); ++j) {
             for (int i = 0; i < C.m(); ++i) {
-                dataC[i + j * strideC] = dataM1[i + j * strideM1] + dataM2[i + j * strideM2];
+                dataC[i + static_cast<size_t>(j) * strideC] = dataM1[i + static_cast<size_t>(j) * strideM1] + dataM2[i + static_cast<size_t>(j) * strideM2];
             }
         }
     }
@@ -305,7 +305,7 @@ void M_Add2(Matrix<Scalar>& M1, Matrix<Scalar>& M2, Matrix<Scalar>& C, double x,
 #endif
         for (int j = 0; j < C.n(); ++j) {
             for (int i = 0; i < C.m(); ++i) {
-                dataC[i + j * strideC] = dataM1[i + j * strideM1] + dataM2[i + j * strideM2] + beta * dataC[i + j * strideC];
+                dataC[i + static_cast<size_t>(j) * strideC] = dataM1[i + static_cast<size_t>(j) * strideM1] + dataM2[i + static_cast<size_t>(j) * strideM2] + beta * dataC[i + static_cast<size_t>(j) * strideC];
             }
         }
     } else {
@@ -314,7 +314,7 @@ void M_Add2(Matrix<Scalar>& M1, Matrix<Scalar>& M2, Matrix<Scalar>& C, double x,
 #endif
         for (int j = 0; j < C.n(); ++j) {
             for (int i = 0; i < C.m(); ++i) {
-                dataC[i + j * strideC] = dataM1[i + j * strideM1] + dataM2[i + j * strideM2];
+                dataC[i + static_cast<size_t>(j) * strideC] = dataM1[i + static_cast<size_t>(j) * strideM1] + dataM2[i + static_cast<size_t>(j) * strideM2];
             }
         }
     }
@@ -334,7 +334,7 @@ void M_Add3(Matrix<Scalar>& M1, Matrix<Scalar>& M2, Matrix<Scalar>& C, double x,
 #endif
         for (int j = 0; j < C.n(); ++j) {
             for (int i = 0; i < C.m(); ++i) {
-                dataC[i + j * strideC] = dataM1[i + j * strideM1] + dataM2[i + j * strideM2] + beta * dataC[i + j * strideC];
+                dataC[i + static_cast<size_t>(j) * strideC] = dataM1[i + static_cast<size_t>(j) * strideM1] + dataM2[i + static_cast<size_t>(j) * strideM2] + beta * dataC[i + static_cast<size_t>(j) * strideC];
             }
         }
     } else {
@@ -343,7 +343,7 @@ void M_Add3(Matrix<Scalar>& M1, Matrix<Scalar>& M2, Matrix<Scalar>& C, double x,
 #endif
         for (int j = 0; j < C.n(); ++j) {
             for (int i = 0; i < C.m(); ++i) {
-                dataC[i + j * strideC] = dataM1[i + j * strideM1] + dataM2[i + j * strideM2];
+                dataC[i + static_cast<size_t>(j) * strideC] = dataM1[i + static_cast<size_t>(j) * strideM1] + dataM2[i + static_cast<size_t>(j) * strideM2];
             }
         }
     }
@@ -363,7 +363,7 @@ void M_Add4(Matrix<Scalar>& M1, Matrix<Scalar>& M2, Matrix<Scalar>& C, double x,
 #endif
         for (int j = 0; j < C.n(); ++j) {
             for (int i = 0; i < C.m(); ++i) {
-                dataC[i + j * strideC] = dataM1[i + j * strideM1] + dataM2[i + j * strideM2] + beta * dataC[i + j * strideC];
+                dataC[i + static_cast<size_t>(j) * strideC] = dataM1[i + static_cast<size_t>(j) * strideM1] + dataM2[i + static_cast<size_t>(j) * strideM2] + beta * dataC[i + static_cast<size_t>(j) * strideC];
             }
         }
     } else {
@@ -372,7 +372,7 @@ void M_Add4(Matrix<Scalar>& M1, Matrix<Scalar>& M2, Matrix<Scalar>& C, double x,
 #endif
         for (int j = 0; j < C.n(); ++j) {
             for (int i = 0; i < C.m(); ++i) {
-                dataC[i + j * strideC] = dataM1[i + j * strideM1] + dataM2[i + j * strideM2];
+                dataC[i + static_cast<size_t>(j) * strideC] = dataM1[i + static_cast<size_t>(j) * strideM1] + dataM2[i + static_cast<size_t>(j) * strideM2];
             }
         }
     }
